% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_docker.R
\name{run_docker_app}
\alias{run_docker_app}
\title{Run the GeoPopMap app inside Docker}
\usage{
run_docker_app(port = 3838, image = "geopopmap_docker_img")
}
\arguments{
\item{port}{The port to expose (default: 3838).}

\item{image}{The name of the Docker image to run.}
}
\description{
This function runs the Docker container that hosts the app.
It will open the app in the default browser.
Requires Docker to be installed.
}
